Separados arquivos de assets das duas versões do censo
