+++ hydra/jobs.py
@@ -21,7 +21,7 @@ geosampa_job = define_asset_job(
 
 censo_job = define_asset_job(
     'censo_job',
-    selection=AssetSelection.groups('censo_bronze')
+    selection=AssetSelection.groups('censo_2010_bronze')
 )
 
 censo_2022_job = define_asset_job(
